Replaced old VonMisesHMM class with new one
