+++ msmbuilder/hmm/__init__.py
@@ -1,3 +1,3 @@
 from __future__ import absolute_import
 from .gaussian import GaussianHMM
-from .vmhmm import VonMisesHMM
+from .vonmises import VonMisesHMM
