--- msmbuilder/hmm/vonmises/spleval.h
@@ -1,6 +1,0 @@
-#ifndef _SPLEVAL_H_
-#define _SPLEVAL_H_
-
-double evaluateSpline(const double* x, const double* y, const double* deriv, long n, double t);
-
-#endif
